requirements.txt
Telethon
python-dotenv
TgCrypto
